+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Latest changed
 
+* publish vehicle info for every vehicle (not only for ego-vehicle)
+* rename CarlaEgoVehicleInfo to CarlaVehicleInfo
+* rename CarlaEgoVehicleInfoWheel to CarlaVehicleInfoWheel
 * tf `child_frame_id` is no similar to topic-prefix (added `/carla`-prefix) 
 * rename carla_ros_bridge_msgs to carla_msgs
 * remove 'challenge' mode